+++ App/controllers/staff.py
@@ -1,18 +1,16 @@
-from App.controllers.user import get_staff
-from App.models import Staff, Student, Review, Karma
 from App.database import db
+from App.models import Staff, Review, Admin
 
-def create_review(staffID, studentID, is_positive, comment):
-    staff = get_staff(staffID)
-    student = db.session.query(Student).get(studentID)
-    
-    if staff and student:
-        review = staff.createReview(student,is_positive, comment)
-        return review
-    return None
+
+def create_review(staffID, studentID, isPositive, comment):
+  staff = get_staff(staffID)
+  review = staff.createReview(studentID, isPositive, comment)
+  if review:
+      return review
+  return False
 
 def search_students_searchTerm(staff, searchTerm):
-    students = staff.searchStudent(searchTerm)
-    if students:
-      return students
-    return None
+  students = staff.searchStudent(searchTerm)
+  if students:
+    return students
+  return None
